Add ngx-cookie-service for token management and refactor related services
