+++ frontend/package.json
@@ -27,6 +27,7 @@
     "apollo-angular": "^8.0.0",
     "express": "^4.21.2",
     "graphql": "^16.9.0",
+    "ngx-cookie-service": "^18.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.10"
