--- frontend/src/app/core/services/auth-guard.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AuthGuard } from './auth-guard.service';
-
-describe('AuthGuardService', () => {
-  let service: AuthGuard;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AuthGuard);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
