--- frontend/src/app/core/services/auth-interceptor.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AuthInterceptorService } from './auth-interceptor.service';
-
-describe('AuthInterceptorService', () => {
-  let service: AuthInterceptorService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AuthInterceptorService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
